requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cmd2==0.8.9
contextlib2==0.5.5
cryptography==2.6.1
defusedxml==0.6.0
enum34==1.1.6
idna==2.8
ipaddress==1.0.22
iso8601==0.1.12
jira==2.0.0
jira-sync==1.0
oauthlib==3.0.1
passlib==1.7.1
pathlib2==2.3.3
pbr==5.1.3
prettytable==0.7.2
prompter==0.3.10
pyaml==19.4.1
pycparser==2.19
PyJWT==1.7.1
pykerberos==1.2.1
pyparsing==2.4.0
pyperclip==1.7.0
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1
requests==2.21.0
requests-kerberos==0.12.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
scandir==1.10.0
six==1.12.0
subprocess32==3.5.3
undecorated==0.3.0
urllib3==1.24.2
wcwidth==0.1.7
yamlordereddictloader==0.4.0
python-redmine==2.2.1
